GCC: Use arguments with same type for std::max
Otherwise GCC produces errors like:
../../base/allocator/partition_allocator/partition_alloc_forward.h:32:38:
error: no matching function for call to 'max(long unsigned int, int)'

Bug: 819294
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3309177
Commit-Queue: Ivan Murashov <ivan.murashov@lge.com>
Cr-Commit-Position: refs/heads/main@{#946887}

+++ base/allocator/partition_allocator/partition_alloc_forward.h
@@ -28,7 +28,8 @@ namespace base {
 // platforms, as Chrome's requirement is C++14 as of 2020.
 #if defined(__STDCPP_DEFAULT_NEW_ALIGNMENT__)
 constexpr size_t kAlignment =
-    std::max(alignof(max_align_t), __STDCPP_DEFAULT_NEW_ALIGNMENT__);
+    std::max(alignof(max_align_t),
+             static_cast<size_t>(__STDCPP_DEFAULT_NEW_ALIGNMENT__));
 #else
 constexpr size_t kAlignment = alignof(max_align_t);
 #endif
